scripts/news/requirements.txt
feedparser>=6.0.0
requests>=2.31.0
openai>=1.0.0
psycopg2-binary>=2.9.0
yfinance>=0.2.0
